arm: tegra: usb_phy: Fix T20, error path
Commit 42274abc973de8565531fe56fae475c162cbfc84 moved the code to set up
the vdd_vbus_typea_usb regulator from T30-specific code to shared code.
This caused T20 to fail to boot, because it also didn't handle the error
case properly (it would notice the error and jump to the error handling
code, but never set the error code so the function would return 0).

Original-Change-Id: I926c0d0c716d409fee0188bc1f766631767a5e3d
Signed-off-by: Robert Morell <rmorell@nvidia.com>
Reviewed-on: http://git-master/r/17978
Reviewed-by: Scott Williams <scwilliams@nvidia.com>
Original-Change-Id: Ie039a639241021afef4774b46c6d87f852715eb7

Rebase-Id: R23ca8928d8de7952e117410b6a281169159c4ec5