media: video: tegra: restore 1264x704 on ov5650
Restore the 1264x704 mode setting, it was erroneously removed
during an integrate from another branch.  Restore the 1920x1088
mode setting I2C sequence, which was removed during the same integrate.
It is not currently used, but we want the sequence there in case we
decide to switch back to it in the future.

Bug 786227

Original-Change-Id: Ia5154f73f8347aae45cb769645460cb49df27b55
Reviewed-on: http://git-master/r/20438
Reviewed-by: Nathan Lord <nlord@nvidia.com>
Tested-by: Nathan Lord <nlord@nvidia.com>
Reviewed-by: Qinggang Zhou <qzhou@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>

Rebase-Id: R59f94a617a3b40b4cfd2bce4c49da937ed6c6df8