ehci:hub: Fix for USB suspend hang
When host_pc is enabled ehci_bus_suspend is disabling the
usb phy by setting the PHCD bit in the HOST_PC register.
Due to this system is hanging during the bus_suspend procedure.
This PHCD bit should be set after completing the USB suspend
procedure. This PHCD bit is set as part of the phy suspend.
Fixing this by removing the code for disabling PHCD bit in
the suspend path.

Bug 793834

Original-Change-Id: I02df4cb9f8ec32c253709ea921f4da5041461f4d
Reviewed-on: http://git-master/r/21492
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Tested-by: Bitan Biswas <bbiswas@nvidia.com>
Tested-by: Suresh Mangipudi <smangipudi@nvidia.com>
Reviewed-by: Hanumanth Venkateswa Moganty <vmoganty@nvidia.com>
Original-Change-Id: Ied6b9e26faf063c956b406ea1d9e1ac50b4727dc

Rebase-Id: Rd49c371c8e2b75e437c69abecaac61d2033ff39f