ARM: tegra: ventana: Fixed backlight toggling.
Ventana has both an internal panel (on fb0) and external HDMI (on
fb1). One of the devices might be blanked while another is unblanked.
In order for the backlight blanking logic to know whether to turn off
the backlight for a particular framebuffer's blanking notification, it
needs to be able to check if a given framebuffer device corresponds to
the backlight.

This implement the check_fb pwm_backlight hook to make the internal
panel's backlight only toggle on fb0 blank changes.

Making equivalent commit from http://git-master/r/21717

Fixed bug 798863

Original-Change-Id: I6465ec40720dc24410808723439277a20c7f764d
Reviewed-on: http://git-master/r/23101
Reviewed-on: http://git-master/r/23726
Tested-by: Gaurav Sarode <gsarode@nvidia.com>
Reviewed-by: Mayuresh Kulkarni <mkulkarni@nvidia.com>
Reviewed-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>

Rebase-Id: R42c55c886e4c7aee975278eb8569cafb79c06da8