media: video: tegra: set camera emc clk preference
Camera requires the emc clk to be a reasonable value, as sensors with
high framerates and resolutions will stress memory bandwidth.  Set to
150 MHz for now.  It will likely need to be fine-tuned and possibly
increased as sensors with higher framerates and resolutions come onto
the market, but for power considerations it should only be bumped
higher if camera really needs the extra throughput.

Bug 913674

Change-Id: I3e900d1a3f468dd0181c42f5a47629fd3b9551b9
Reviewed-on: http://git-master/r/73891
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Nathan Lord <nlord@nvidia.com>
Tested-by: Nathan Lord <nlord@nvidia.com>
Reviewed-by: Jihoon Bang <jbang@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-on: http://git-master/r/75141
Reviewed-by: Varun Wadekar <vwadekar@nvidia.com>
Tested-by: Varun Wadekar <vwadekar@nvidia.com>