ARM: tegra: dma: Pause dma before configuring dma for next transfer
The apb dma supports next transfer configuration before current
on-going transfer completes in continuous mode.
The new configuration require multiple register programming and need
to be done before last burst completed. Just after last transfer
completes, the new configure data is reloaded.
To make this configuration atomic, pausing the dma during configuration
so that last burst should not be happen and dma engine should not re-load
new configuration when it was on a way.

bug 937142

Change-Id: I15b62394df10c97ca21c0d7905fedc7e7c2872b7
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/84292
Tested-by: Pradeep Goudagunta <pgoudagunta@nvidia.com>
Tested-by: Bandi Krishna Chaitanya <bandik@nvidia.com>
Reviewed-by: Stephen Warren <swarren@nvidia.com>