ALSA:hda: add alsa control for query of device
alsa control is added to facilitate querying dts decode capability of
connected device. dts decode capability of the connected device is
updated in the ELD buffer. That information is updated in the
ALSA control structure. In addition to that, the code is amended to
handle other pass through decoder mode support.

Bug 943017

Change-Id: If8da7dfb24be3b86592191f5586b70492282b438
Signed-off-by: Sayak Ghosh Choudhury <sayakc@nvidia.com>
Reviewed-on: http://git-master/r/91813
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Scott Peterson <speterson@nvidia.com>