usb: otg: clear the VBUS INT bit when ID status is detected
The function tegra_otg_set_peripheral() is called on probing stage
to apply the correct otg state on boot.
However, when the otg cable is plugged in on boot, the VBUS flags
are not correct leading to misinterpreation of states.
For now, we will clear the VBUS INT flag to make the state correct.

[Issue 6464161] USB host can not work if the device boots up with the
USB host cable connected

Change-Id: Iab6d9529fe9e7aa5e5f0552bc4fcee6b2a4e1c64