arm: tegra: usb_phy: disable PMC mode for USB1
Both ehci host and udc device are registered on USB1. Thus we
have both device and host mode phy for USB1.
In this case, when we back from LP0, utmi_phy_power_on()
were called twice, one for device mode and the other for host mode.
For the device mode, it's called by fsl_udc_resume() then the phy
is turned off right after the udc driver detected the id pin is low
(we are working on host mode). However, the PMC is enabled for
usb1 now by calling utmip_powerdown_pmc_wake_detect(), which
is call by utmi_phy_power_off().

PMC mode needs to be disabled to switch the pin control back to the
host controller.

bug 984119

Change-Id: I7bcce88cf5f42e0cbf3ee2d0cc24bbcadf3d51e9
Signed-off-by: Ken Chang <kenc@nvidia.com>