tegra_hotplug: new algorithm
Every 0.1 second it polls for load. if the load is higher than the "threshold"
(configurable via first_level), it scores 1 point. If the load is lower than
the "threshold", it takes away 1 point. On the scoreboard, if the points reach
10 --> 4 cores -> polls every 0.25 second from now on. Whenever the points
on the scoreboard drops below 10 --> 1 core operation --> polling
(frequency of calculating scores) back to 0.1 second.

So basically the hotplug driver has either 1 core or 4 cores. And 2 cores are
up by touch boosts.

"lighter games" will have 4 cores too because the load is not calculated from
all the cores. Instead the new driver only takes the load from cpu0
(most games, esp light ones focus on stressing cpu0) so there will be a higher
chance to reach 10 points on the scoreboard --> 4 cores --> points dont drop
because of constant gaming --> constant 4 cores ->> ultimate fps while gaming.

My theory is that whenever you game, you dont care about battery and so the
kernel gives best fps. Out of gaming? The prevention of unnecessary cores
jumping will save more juice!

Also for anyone worrying about the scoreboard having too many points, the
maximum points for the cpus to score is 50.

Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>